+++ test/device/test_device_configuration.py
@@ -185,6 +185,17 @@ def test_close_defined_yaml_device(moler_config, device_factory):
     assert device_2 != device_org
 
 
+def test_can_remove_device_twice(moler_config, device_factory):
+    conn_config = os.path.join(os.path.dirname(__file__), os.pardir, "resources", "device_config.yml")
+    moler_config.load_config(config=conn_config, config_type='yaml')
+
+    device_org_name = 'UNIX_LOCAL'
+    device_org = device_factory.get_device(name=device_org_name)
+    assert device_org is not None
+    device_org.remove()
+    device_org.remove()
+
+
 def test_clone_device_from_cloned_device(moler_config, device_factory):
     conn_config = os.path.join(os.path.dirname(__file__), os.pardir, "resources", "device_config.yml")
     moler_config.load_config(config=conn_config, config_type='yaml')
@@ -626,6 +637,45 @@ def test_cannot_load_configuration_with_the_same_named_device_loaded_from_anothe
     with pytest.raises(WrongUsage) as err:
         moler_config.load_config(config=new_conn_config, config_type='dict')
     assert "but now requested with SM params" in str(err.value)
+
+
+def test_load_device_from_config(moler_config, device_factory):
+    config1 = {
+        'LOGGER': {
+            'PATH': '/tmp/',
+            'RAW_LOG': True,
+            'DATE_FORMAT': '%d %H:%M:%S'
+        },
+        'DEVICES': {
+            'UNIX_LOCAL': {
+                'DEVICE_CLASS': 'moler.device.unixlocal.UnixLocal',
+                'INITIAL_STATE': 'UNIX_LOCAL'
+            }
+        }
+    }
+    config2 = {
+        'LOGGER': {
+            'PATH': '/tmp/',
+            'RAW_LOG': True,
+            'DATE_FORMAT': '%d %H:%M:%S'
+        },
+        'DEVICES': {
+            'UNIX_LOCAL2': {
+                'DEVICE_CLASS': 'moler.device.unixlocal.UnixLocal',
+                'INITIAL_STATE': 'UNIX_LOCAL'
+            }
+        }
+    }
+    moler_config.load_config(config1, None, 'dict')
+    dev1_prev = device_factory.get_device("UNIX_LOCAL")
+    with pytest.raises(KeyError):
+        device_factory.get_device("UNIX_LOCAL2")
+
+    moler_config.load_device_from_config(config2)
+    dev = device_factory.get_device('UNIX_LOCAL2')
+    assert dev is not None
+    dev1_after = device_factory.get_device('UNIX_LOCAL')
+    assert dev1_prev == dev1_after
 
 
 def test_create_device_without_hops():